# api/internal/api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package api

import (
	"context"
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes bearerAuthContextKey = "bearerAuth.Scopes"
)

// bearerAuthContextKey is the context key for bearerAuth security scheme
type bearerAuthContextKey string

// AuthLoginJSONBody defines parameters for AuthLogin.
type AuthLoginJSONBody struct {
	Email    *string `json:"email,omitempty"`
	Password *string `json:"password,omitempty"`
}

// ListEmailsParams defines parameters for ListEmails.
type ListEmailsParams struct {
	Page    *int `form:"page,omitempty" json:"page,omitempty"`
	PerPage *int `form:"per_page,omitempty" json:"per_page,omitempty"`
}

// MarkReadJSONBody defines parameters for MarkRead.
type MarkReadJSONBody struct {
	Read *bool `json:"read,omitempty"`
}

// CreateNewSmtpTokenJSONBody defines parameters for CreateNewSmtpToken.
type CreateNewSmtpTokenJSONBody struct {
	Description *string    `json:"description,omitempty"`
	ExpiresAt   *time.Time `json:"expires_at,omitempty"`
}

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody AuthLoginJSONBody

// MarkReadJSONRequestBody defines body for MarkRead for application/json ContentType.
type MarkReadJSONRequestBody MarkReadJSONBody

// CreateNewSmtpTokenJSONRequestBody defines body for CreateNewSmtpToken for application/json ContentType.
type CreateNewSmtpTokenJSONRequestBody CreateNewSmtpTokenJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/auth/login)
	AuthLogin(w http.ResponseWriter, r *http.Request)

	// (GET /api/healthz)
	Health(w http.ResponseWriter, r *http.Request)

	// (GET /api/organizations/{orgId}/emails)
	ListEmails(w http.ResponseWriter, r *http.Request, orgId string, params ListEmailsParams)

	// (GET /api/organizations/{orgId}/emails/unread-count)
	UnreadCount(w http.ResponseWriter, r *http.Request, orgId string)

	// (DELETE /api/organizations/{orgId}/emails/{id})
	DeleteEmail(w http.ResponseWriter, r *http.Request, orgId string, id string)

	// (GET /api/organizations/{orgId}/emails/{id})
	GetEmail(w http.ResponseWriter, r *http.Request, orgId string, id string)

	// (PUT /api/organizations/{orgId}/emails/{id}/read)
	MarkRead(w http.ResponseWriter, r *http.Request, orgId string, id string)

	// (GET /api/organizations/{orgId}/smtp/tokens)
	GetTokenList(w http.ResponseWriter, r *http.Request, orgId string)

	// (POST /api/organizations/{orgId}/smtp/tokens)
	CreateNewSmtpToken(w http.ResponseWriter, r *http.Request, orgId string)

	// (DELETE /api/organizations/{orgId}/smtp/tokens/{tokenId})
	DeleteSmtpToken(w http.ResponseWriter, r *http.Request, orgId string, tokenId string)

	// (GET /api/organizations/{orgId}/smtp/tokens/{tokenId})
	GetSmtpToken(w http.ResponseWriter, r *http.Request, orgId string, tokenId string)

	// (GET /api/users/me/organizations)
	ListOrganizations(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (POST /api/auth/login)
func (_ Unimplemented) AuthLogin(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/healthz)
func (_ Unimplemented) Health(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/organizations/{orgId}/emails)
func (_ Unimplemented) ListEmails(w http.ResponseWriter, r *http.Request, orgId string, params ListEmailsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/organizations/{orgId}/emails/unread-count)
func (_ Unimplemented) UnreadCount(w http.ResponseWriter, r *http.Request, orgId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (DELETE /api/organizations/{orgId}/emails/{id})
func (_ Unimplemented) DeleteEmail(w http.ResponseWriter, r *http.Request, orgId string, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/organizations/{orgId}/emails/{id})
func (_ Unimplemented) GetEmail(w http.ResponseWriter, r *http.Request, orgId string, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /api/organizations/{orgId}/emails/{id}/read)
func (_ Unimplemented) MarkRead(w http.ResponseWriter, r *http.Request, orgId string, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/organizations/{orgId}/smtp/tokens)
func (_ Unimplemented) GetTokenList(w http.ResponseWriter, r *http.Request, orgId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/organizations/{orgId}/smtp/tokens)
func (_ Unimplemented) CreateNewSmtpToken(w http.ResponseWriter, r *http.Request, orgId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (DELETE /api/organizations/{orgId}/smtp/tokens/{tokenId})
func (_ Unimplemented) DeleteSmtpToken(w http.ResponseWriter, r *http.Request, orgId string, tokenId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/organizations/{orgId}/smtp/tokens/{tokenId})
func (_ Unimplemented) GetSmtpToken(w http.ResponseWriter, r *http.Request, orgId string, tokenId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/users/me/organizations)
func (_ Unimplemented) ListOrganizations(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// AuthLogin operation middleware
func (siw *ServerInterfaceWrapper) AuthLogin(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Health operation middleware
func (siw *ServerInterfaceWrapper) Health(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Health(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListEmails operation middleware
func (siw *ServerInterfaceWrapper) ListEmails(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ListEmailsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", r.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "page"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "per_page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "per_page", r.URL.Query(), &params.PerPage, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "per_page"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "per_page", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListEmails(w, r, orgId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UnreadCount operation middleware
func (siw *ServerInterfaceWrapper) UnreadCount(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UnreadCount(w, r, orgId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteEmail operation middleware
func (siw *ServerInterfaceWrapper) DeleteEmail(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteEmail(w, r, orgId, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetEmail operation middleware
func (siw *ServerInterfaceWrapper) GetEmail(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetEmail(w, r, orgId, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MarkRead operation middleware
func (siw *ServerInterfaceWrapper) MarkRead(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MarkRead(w, r, orgId, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTokenList operation middleware
func (siw *ServerInterfaceWrapper) GetTokenList(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTokenList(w, r, orgId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateNewSmtpToken operation middleware
func (siw *ServerInterfaceWrapper) CreateNewSmtpToken(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateNewSmtpToken(w, r, orgId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteSmtpToken operation middleware
func (siw *ServerInterfaceWrapper) DeleteSmtpToken(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	// ------------- Path parameter "tokenId" -------------
	var tokenId string

	err = runtime.BindStyledParameterWithOptions("simple", "tokenId", chi.URLParam(r, "tokenId"), &tokenId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "tokenId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteSmtpToken(w, r, orgId, tokenId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSmtpToken operation middleware
func (siw *ServerInterfaceWrapper) GetSmtpToken(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orgId" -------------
	var orgId string

	err = runtime.BindStyledParameterWithOptions("simple", "orgId", chi.URLParam(r, "orgId"), &orgId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orgId", Err: err})
		return
	}

	// ------------- Path parameter "tokenId" -------------
	var tokenId string

	err = runtime.BindStyledParameterWithOptions("simple", "tokenId", chi.URLParam(r, "tokenId"), &tokenId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "tokenId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSmtpToken(w, r, orgId, tokenId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListOrganizations operation middleware
func (siw *ServerInterfaceWrapper) ListOrganizations(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListOrganizations(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/auth/login", wrapper.AuthLogin)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/healthz", wrapper.Health)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/organizations/{orgId}/emails", wrapper.ListEmails)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/organizations/{orgId}/emails/unread-count", wrapper.UnreadCount)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/organizations/{orgId}/emails/{id}", wrapper.DeleteEmail)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/organizations/{orgId}/emails/{id}", wrapper.GetEmail)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/api/organizations/{orgId}/emails/{id}/read", wrapper.MarkRead)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/organizations/{orgId}/smtp/tokens", wrapper.GetTokenList)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/organizations/{orgId}/smtp/tokens", wrapper.CreateNewSmtpToken)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/organizations/{orgId}/smtp/tokens/{tokenId}", wrapper.DeleteSmtpToken)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/organizations/{orgId}/smtp/tokens/{tokenId}", wrapper.GetSmtpToken)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/users/me/organizations", wrapper.ListOrganizations)
	})

	return r
}